Carsten, is my HEAD kdelibs too old now or is this indeed ambiguous?
Either way, I needed this to fix compile:

libkopete/ui/kopetefileconfirmdialog.cpp:43: call of overloaded `formatNumber(long unsigned int)' is ambiguous
/usr/local/kde3/include/klocale.h:371: candidates are: class QString KLocale::formatNumber(double, int = -1) const
/usr/local/kde3/include/klocale.h:384:                 class QString KLocale::formatNumber(long int) const

svn path=/trunk/kdenetwork/kopete/; revision=212531

+++ libkopete/ui/kopetefileconfirmdialog.cpp
@@ -40,7 +40,7 @@ KopeteFileConfirmDialog::KopeteFileConfirmDialog(const KopeteFileTransferInfo &i
 	m_view=new FileConfirmBase(this, "FileConfirmView");
 	m_view->m_from->setText( info.contact()->metaContact()->displayName() + QString::fromLatin1( " <" ) +
 		info.contact()->contactId() + QString::fromLatin1( "> " ) );
-	m_view->m_size->setText( KGlobal::locale()->formatNumber( info.size() ) );
+	m_view->m_size->setText( KGlobal::locale()->formatNumber( long( info.size() ) ) );
 	m_view->m_description->setText( description );
 	m_view->m_filename->setText( info.file() );
 	m_view->m_saveto->setText( QDir::homeDirPath() + QString::fromLatin1( "/" ) + info.file() );
